+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added Projects resource
+- Added Expense Categories resource
 - Added list sort builder
 - Updated list models with new `pages()` function to standardize pagination with different models